Removing htmlPrinter abstraction
