--- src/Behat/Behat/Output/Node/Printer/Html/AbstractHtmlPrinter.php
@@ -1,48 +1,0 @@
-<?php
-
-/*
- * This file is part of the Behat.
- * (c) Konstantin Kudryashov <ever.zet@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Behat\Behat\Output\Node\Printer\Html;
-
-use Behat\Behat\Output\Node\Printer\Helper\HtmlPrinter;
-use Behat\Testwork\Output\Printer\OutputPrinter;
-
-/**
- * To retrieve (html printer) helper .
- *
- * @author Ali Bahman <abn@webit4.me>
- */
-abstract class AbstractHtmlPrinter
-{
-
-    /**
-     * @var HtmlPrinter
-     */
-    private $htmlPrinter;
-
-    /**
-     * @param OutputPrinter $printer
-     * @return HtmlPrinter
-     */
-    protected function getHtmlPrinter(OutputPrinter $printer)
-    {
-        $this->htmlPrinter->setOutputPrinter($printer);
-        return $this->htmlPrinter;
-    }
-
-    /**
-     * @param OutputPrinter $printer
-     * @return HtmlPrinter
-     */
-    protected function setHtmlPrinter(HtmlPrinter $htmlPrinter)
-    {
-        $this->htmlPrinter = $htmlPrinter;
-        return $this;
-    }
-}
